refactor: update util class name
